requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
motor==3.3.2
pymongo==4.6.1
openai==1.12.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
gunicorn==21.2.0
nltk==3.8.1 